+++ src/titiler/mosaic/tests/test_factory.py
@@ -5,7 +5,7 @@ import tempfile
 from contextlib import contextmanager
 from dataclasses import dataclass
 from io import BytesIO
-from typing import Callable, Optional
+from typing import Optional
 
 import attr
 import numpy
@@ -247,40 +247,6 @@ def _multiply_by_two(data, mask):
     mask.fill(255)
     data = data * 2
     return data, mask
-
-
-@dataclass
-class ReaderParams(DefaultDependency):
-    """Backend options to overwrite min/max zoom."""
-
-    post_process: Callable = _multiply_by_two
-
-
-def test_MosaicTilerFactory_ReaderParams():
-    """Test MosaicTilerFactory factory with Reader dependency."""
-    mosaic = MosaicTilerFactory(router_prefix="/mosaic")
-    mosaic_two = MosaicTilerFactory(
-        reader_dependency=ReaderParams, router_prefix="/mosaic_two"
-    )
-
-    app = FastAPI()
-    app.include_router(mosaic.router, prefix="/mosaic")
-    app.include_router(mosaic_two.router, prefix="/mosaic_two")
-    client = TestClient(app)
-
-    with tmpmosaic() as mosaic_file:
-        response = client.get(
-            "/mosaic/point/-74.53125,45.9956935",
-            params={"url": mosaic_file},
-        )
-        value = response.json()["values"][0][1][0]
-
-        response = client.get(
-            "/mosaic_two/point/-74.53125,45.9956935",
-            params={"url": mosaic_file},
-        )
-        value_two = response.json()["values"][0][1][0]
-        assert value_two == 2 * value
 
 
 def test_MosaicTilerFactory_PixelSelectionParams():